pyproject.toml
[project]
name = "trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "feedparser>=6.0.12",
    "google-generativeai>=0.8.5",
    "holidays>=0.85",
    "munch>=4.0.0",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "tenacity>=9.1.2",
    "trading-ig>=0.0.22",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
